requirements.txt
asgiref==3.5.0
click==8.0.4       
colorama==0.4.4    
cycler==0.11.0     
Flask==2.0.3       
fonttools==4.31.2  
h11==0.13.0        
itsdangerous==2.1.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.4.0
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
PySastrawi==1.2.0
python-dateutil==2.8.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
Werkzeug==2.0.3